python-ai-service/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
openai>=1.10.0
pydantic>=2.5.0
python-dotenv>=1.0.0
httpx>=0.26.0
python-multipart>=0.0.6
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-core>=0.1.0
pinecone>=4.0.0
